create new DK parser for exchanges Renamed the old parser to DK_old.py and archived it.
